backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grafana-dashboard-generator"
version = "0.1.0"
description = "AI-Driven Grafana Dashboard Generator"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "AI Assistant", email = "example@example.com"},
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "langchain>=0.1.12",
    "langchain-openai>=0.1.5",
    "langchain-anthropic>=0.1.5",
    "langgraph>=0.1.5",
    "jsonschema>=4.21.0",
    "python-dotenv>=1.0.0",
    "openai>=1.12.0",
    "anthropic>=0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.9.1",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "ruff>=0.0.290",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.hatch.build.targets.wheel]
packages = ["app"]
